Changes
Added objects:
Radio Station - a class which starts and stops the audio player.
Player - structure which sets up the UI player as an AVAudioPlayer.
Track - structure which takes the JSON radio station data and parses it
to var
Networking Class:
DataManager - Gets the streaming URL and casts it as a string, along
with song metadata and error handlers.
Support :
RadioSettings - sets up the variables that the radio player will use.
Views:
NowPlayingViewController - Started linking the objects to the sheet,
stopped at the constraints.